refactor(openapi): move PaymentRuleResponse and OrisEventSummary to top-level classes
Both were nested classes requiring a schemaMappings redirect since a bare
OpenAPI schema name never auto-resolves to a nested class. Moving them to
top-level files in the same package lets the generator resolve them directly,
eliminating both mappings entries.
