+++ backend/src/main/java/com/klabis/oris/OrisController.java
@@ -65,7 +65,4 @@ public class OrisController implements OrisImportApi {
 
         return ResponseEntity.ok(events);
     }
-
-    public record OrisEventSummary(int id, String name, LocalDate date, String location, String organizer) {
-    }
 }
