+++ backend/src/main/java/com/klabis/oris/OrisEventSummary.java
@@ -1,0 +1,6 @@
+package com.klabis.oris;
+
+import java.time.LocalDate;
+
+public record OrisEventSummary(int id, String name, LocalDate date, String location, String organizer) {
+}
